requirements.txt
pyserial>=3.5
numpy>=1.19.0
Pillow>=8.0.0